Reorganize codebase with feature-based directory structure
Move root-level classes into organized directories:
- src/core/: Domain models (Author, Publication, Filter, PublicationSearch)
- src/lib/: Infrastructure utilities (Cache, Keys, Util, FpsTracker)

Update all import statements across 60+ files including components,
stores, composables, services, and test files. Update CLAUDE.md
documentation with new directory organization.

All tests passing (29 test files, 400 tests). Build successful.
