+++ src/components/NetworkPerformanceMonitor.vue
@@ -27,7 +27,7 @@
 
 <script setup>
 import { ref } from 'vue';
-import { FpsTracker } from '@/FpsTracker.js';
+import { FpsTracker } from '@/lib/FpsTracker.js';
 
 // Props definition with defaults  
 defineProps({
