+++ src/composables/useAppState.js
@@ -4,7 +4,7 @@ import { useInterfaceStore } from '@/stores/interface.js'
 import { useAuthorStore } from '@/stores/author.js'
 import { useQueueStore } from '@/stores/queue.js'
 import { PAGINATION } from '@/constants/config.js'
-import { clearCache as clearCacheUtil } from '@/Cache.js'
+import { clearCache as clearCacheUtil } from '@/lib/Cache.js'
 import { SuggestionService } from '@/services/SuggestionService.js'
 
 export function useAppState() {
