+++ src/core/Author.js
@@ -1,4 +1,4 @@
-import { SCORING } from './constants/config.js';
+import { SCORING } from '../constants/config.js';
 
 export default class Author {
