+++ src/core/Filter.js
