+++ src/lib/FpsTracker.js
