+++ tests/unit/Author.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest'
-import Author from '@/Author.js'
+import Author from '@/core/Author.js'
 
 // Mock constants
 vi.mock('@/constants/publication.js', () => ({
