+++ tests/unit/Filter.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest'
-import Filter from '@/Filter.js'
+import Filter from '@/core/Filter.js'
 
 describe('Filter', () => {
   let filter
