+++ tests/unit/components/PublicationListComponent.test.js
@@ -10,12 +10,12 @@ vi.mock('@/stores/session.js')
 vi.mock('@/stores/interface.js')
 
 // Mock the Cache module to avoid indexedDB issues
-vi.mock('@/Cache.js', () => ({
+vi.mock('@/lib/Cache.js', () => ({
   clearCache: vi.fn()
 }))
 
 // Mock utilities
-vi.mock('@/Util.js', () => ({
+vi.mock('@/lib/Util.js', () => ({
   scrollToTargetAdjusted: vi.fn(),
   shuffle: vi.fn(arr => arr),
   saveAsFile: vi.fn()
