+++ tests/unit/components/QuickAccessBar.test.js
@@ -5,7 +5,7 @@ import QuickAccessBar from '@/components/QuickAccessBar.vue'
 import { useQueueStore } from '@/stores/queue.js'
 
 // Mock scroll utility
-vi.mock('@/Util.js', () => ({
+vi.mock('@/lib/Util.js', () => ({
   scrollToTargetAdjusted: vi.fn()
 }))
